Closes #4901
Now bv_update_projects_list does not remove existing information from brainvisa-cmake files, it only adds new information got by parsing the repository.
To force deleting information about projects or components that are not found in the repository, the user have to use the option -d (delete) but this option should be used only if the user have the access rights on the whole repository.
It is also possible to ask for updating only a subset of projects and components using the options -p (project) and -c (component). The script will parse only the requested projects and components and will keep previous information for others.

For example, if someone creates a new branch in axon, it is possible to call bv_update_projects_list to update only axon information:
bv_update_projects_list -p axon



git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/development/brainvisa-cmake/trunk@49040 613412ce-1361-4b7a-93bf-a3c75af1bae4