Fail bv_maker sources when a Git repository cannot be updated
This should prevent user confusion where someone thinks that their sources are
up-to-date because they ran bv_maker, but some Git repositories are not
updated.

However, this means that bv_maker will stop after fetching the sources and skip
the configure and build steps if any of your Git repositories has diverged. You
are welcome to add a command-line/configuration flag for ignoring Git update
errors if this turns out to be a limitation for your workflow.