bv_maker: hotfix for 'bv_maker sources' on Python 2
bv_maker failed when the Python default string encoding was set to 'ascii'.
This is the case by default, when there is no 'sitecustomize' module. This bug
should not have affected bv_maker when run in an existing BrainVISA build tree
through bv_env_host, because the build tree has such a 'sitecustomize' module
that sets the default encoding to utf-8.