Remove packaging of external libraries
Since we now distribute BrainVISA in containers we can finally
get rid of the hassle of re-packaging all these external libraries.

This commit only removes the packaging code that is specific to
external libraries, the packaging of our own tools should not
be affected (even though we are not using it at the moment).

Closes: https://github.com/brainvisa/brainvisa-cmake/issues/82